# code/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parking foundation — local development",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/test/agent/operations": {
      "post": {
        "summary": "Autonomous Control",
        "operationId": "autonomous_control_api_v1_test_agent_operations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutonomousControl"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Autonomous Status",
        "operationId": "autonomous_status_api_v1_test_agent_operations_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/test/agent/operations/jobs": {
      "get": {
        "summary": "Autonomous Jobs",
        "operationId": "autonomous_jobs_api_v1_test_agent_operations_jobs_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commands/{command_id}/clarify": {
      "post": {
        "summary": "Clarify Command",
        "operationId": "clarify_command_api_v1_commands__command_id__clarify_post",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Command Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandClarification"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commands/{command_id}/plan": {
      "get": {
        "summary": "Command Plan",
        "operationId": "command_plan_api_v1_commands__command_id__plan_get",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Command Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commands/{command_id}/confirm": {
      "post": {
        "summary": "Confirm Command",
        "operationId": "confirm_command_api_v1_commands__command_id__confirm_post",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Command Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutonomousControl": {
        "properties": {
          "run_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Run Id"
          },
          "action": {
            "type": "string",
            "enum": [
              "start",
              "stop",
              "process"
            ],
            "title": "Action"
          },
          "mode": {
            "type": "string",
            "enum": [
              "mock",
              "live"
            ],
            "title": "Mode",
            "default": "mock"
          },
          "scenario": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "s1a",
                  "s1b",
                  "s1c",
                  "s2",
                  "s3"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario"
          },
          "command_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Command Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "action"
        ],
        "title": "AutonomousControl"
      },
      "CancelInput": {
        "properties": {
          "expected_resource_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Resource Version"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_resource_version"
        ],
        "title": "CancelInput"
      },
      "CommandClarification": {
        "properties": {
          "expected_resource_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Resource Version"
          },
          "goal": {
            "type": "string",
            "enum": [
              "closing",
              "zone_notice"
            ],
            "title": "Goal"
          },
          "zone_id": {
            "anyOf": [
              {
                "type": "string",
                "const": "announcement-a"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_resource_version",
          "goal"
        ],
        "title": "CommandClarification"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}